Prompt Engineering - Project 2/nl2cmd/pyproject.toml
[project]
name = "nl2cmd"
version = "0.1.0"
description = "Natural Language to CMD Agent"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "gradio>=4.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
